Split runtime metadata handlers
